[mbed access point] IPv4 address of mbed Access Point
After first boot of mbed access point, it is necessary to
know the IP address in order to commmunicate with it. Present
method is to connect a monitor to see the IP address on the
screen or sniff the lldp packets sent by the mbed access point
with the help of wireshark.

The boot partition of the image is auto-mounted when micro-sd
card is inserted into a computer. With this commit, a file named
ip_address.txt is created in the boot partition, so that user can
get the IP address by simply inserting the micro-sd card into
computer after the first boot of mbed access point.

Signed-off-by: Devaraj Ranganna <devaraj.ranganna@arm.com>